clouder-back/backend/requirements.txt
requests
boto3
pybase64
pyjwt
